Handle the api exceptions

+++ src/main/java/Main.java
@@ -1,14 +1,16 @@
 import com.google.gson.Gson;
 import dao.Sql2oGrandSlamDao;
 import dao.Sql2oTennisPlayerDao;
-import dao.TennisPlayerDao;
+import exceptions.ApiException;
 import models.Country;
 import models.GrandSlam;
 import models.TennisPlayer;
 import org.sql2o.Connection;
 import org.sql2o.Sql2o;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static spark.Spark.*;
 
@@ -40,6 +42,9 @@ public class Main {
             int playerId = Integer.parseInt(request.params("id"));
             String countryName = request.params("countryName");
             TennisPlayer player = tennisPlayerDao.findById(playerId);
+            if (player == null) {
+                throw new ApiException(String.format("No player with id = %d found", playerId), 404);
+            }
             Country country = new Country(countryName);
             tennisPlayerDao.addCountryToPlayer(player, country);
             response.status(201);
@@ -51,12 +56,17 @@ public class Main {
             int playerId = Integer.parseInt(request.params("id"));
             int tournmId = Integer.parseInt(request.params("tournamentId"));
             TennisPlayer player = tennisPlayerDao.findById(playerId);
+            if (player == null) {
+                throw new ApiException(String.format("No player with id = %d found", playerId), 404);
+            }
             GrandSlam tournament = grandSlamDao.findById(tournmId);
+            if (tournament == null) {
+                throw new ApiException(String.format("No tournament with id = %d found", tournmId), 404);
+            }
             tennisPlayerDao.addPlayerToTournament(player, tournament);
             response.status(201);
             return gson.toJson(player);
         });
-
 
         // READ
         // all players
@@ -66,16 +76,22 @@ public class Main {
         });
 
         // single player
-        // TODO: handle an exception
         get("/players/:id", "application/json", (request, response) -> {
             int playerId = Integer.parseInt(request.params("id"));
             TennisPlayer player = tennisPlayerDao.findById(playerId);
+            if (player == null) {
+                throw new ApiException(String.format("No player with id = %d found", playerId), 404);
+            }
             return gson.toJson(player);
         });
 
         // get all tournaments won by a player
         get("/players/:id/tournaments", "application/json", (request, response) -> {
             int playerId = Integer.parseInt(request.params("id"));
+            TennisPlayer player = tennisPlayerDao.findById(playerId);
+            if (player == null) {
+                throw new ApiException(String.format("No player with id = %d found", playerId), 404);
+            }
             List<GrandSlam> tournaments = tennisPlayerDao.getAllTournamentsWonByPlayer(playerId);
             return gson.toJson(tournaments);
         });
@@ -89,9 +105,12 @@ public class Main {
         });
 
         // UPDATE
-        post("/players/:id/update", "applciation/json", (request, response) -> {
+        post("/players/:id/update", "application/json", (request, response) -> {
             int playerId = Integer.parseInt(request.params("id"));
             TennisPlayer player = gson.fromJson(request.body(), TennisPlayer.class);
+            if (player == null) {
+                throw new ApiException(String.format("No player with id = %d found", playerId), 404);
+            }
             tennisPlayerDao.update(playerId, player.getAge(), player.getRanking(), player.getPoints(), player.getTournamentsPlayed());
             response.status(201);
             return gson.toJson(player);
@@ -101,6 +120,10 @@ public class Main {
         // single player
         get("/players/:id/delete", "application/json", (request, response) -> {
             int playerId = Integer.parseInt(request.params("id"));
+            TennisPlayer player = tennisPlayerDao.findById(playerId);
+            if (player == null) {
+                throw new ApiException(String.format("No player with id = %d found", playerId), 404);
+            }
             tennisPlayerDao.deletePlayer(playerId);
             return gson.toJson(playerId);
         });
@@ -113,7 +136,7 @@ public class Main {
 
 
         // TOURNAMENTS
-        //CREATE
+        // CREATE
         post("/tournaments/new", "application/json", (request, response) -> {
             GrandSlam tournament = gson.fromJson(request.body(), GrandSlam.class);
             grandSlamDao.add(tournament);
@@ -131,7 +154,7 @@ public class Main {
             return gson.toJson(tournament);
         });
 
-        //READ
+        // READ
         // all tournaments
         get("/tournaments", "application/json", (request, response) -> {
             List<GrandSlam> tournaments = grandSlamDao.getAllTornaments();
@@ -139,16 +162,22 @@ public class Main {
         });
 
         // single tournament
-        // TODO: handle an exception
         get("/tournaments/:tournId", "application/json", (request, response) -> {
             int tournId = Integer.parseInt(request.params("tournId"));
             GrandSlam tournament = grandSlamDao.findById(tournId);
+            if (tournament == null) {
+                throw new ApiException(String.format("No tournament with id = %d found", tournId), 404);
+            }
             return gson.toJson(tournament);
         });
 
         // get all players won the tournament
-        get("/tournaments/:tournId/players", "applciation/json", (request, response) -> {
+        get("/tournaments/:tournId/players", "application/json", (request, response) -> {
             int tournId = Integer.parseInt(request.params("tournId"));
+            GrandSlam tournament = grandSlamDao.findById(tournId);
+            if (tournament == null) {
+                throw new ApiException(String.format("No tournament with id = %d found", tournId), 404);
+            }
             List<TennisPlayer> players = grandSlamDao.getAllPlayersWonTheTournament(tournId);
             return gson.toJson(players);
         });
@@ -157,6 +186,10 @@ public class Main {
         // single tournament
         get("/tournaments/:tournId/delete", "application/json", (request, response) -> {
             int tournId = Integer.parseInt(request.params("tournId"));
+            GrandSlam tournament = grandSlamDao.findById(tournId);
+            if (tournament == null) {
+                throw new ApiException(String.format("No tournament with id = %d found", tournId), 404);
+            }
             grandSlamDao.deleteTourn(tournId);
             return gson.toJson(tournId);
         });
@@ -165,6 +198,17 @@ public class Main {
         get("/tournaments/delete", "application/json", (request, response) -> {
             grandSlamDao.deleteAllTourn();
             return gson.toJson(grandSlamDao.getAllTornaments().size());
+        });
+
+        // exception
+        exception(ApiException.class, (exception, request, response) -> {
+            ApiException err = (ApiException) exception;
+            Map<String, Object> jsonMap = new HashMap<>();
+            jsonMap.put("status", err.getStatusCode());
+            jsonMap.put("errorMessage", err.getMessage());
+            response.type("application/json");
+            response.status(err.getStatusCode());
+            response.body(gson.toJson(jsonMap));
         });
     }
 }
